files are placed in Scenario Selection Module

+++ SSTSS_GenSim_Modules/Scenario_Selection_Module/simulator_filtering.py
@@ -1,24 +1,47 @@
-import os
-from modules.constants import scenarios_excel_file_name
-
 import pandas as pd
-
-
+from Scenario_Selection_Module import (
+    PRIORITIZED_SCENARIO_GROUPS_US_PATH,
+    PRIORITIZED_SCENARIO_GROUPS_EU_PATH,
+    FILTERED_SCENARIOS_CARLA_PATH,
+    FILTERED_SCENARIOS_GAZEBO_PATH,
+    FILTERED_SCENARIOS_AUDACITY_PATH,
+    FILTERED_SCENARIOS_LGSVL_PATH
+)
 
 def filter_scenarios_based_on_simulator(simulator_name, dataset_name):
-    # Load the Excel file into a pandas DataFrame
-    dataset_name = dataset_name.lower()
 
-    df = pd.read_excel(os.getcwd() + f"/prioritized_scenario_groups_{dataset_name}.xlsx")
+    # Dataset selection
+    dataset_name = dataset_name.upper()
+    if dataset_name == "US":
+        input_file = PRIORITIZED_SCENARIO_GROUPS_US_PATH
+    elif dataset_name == "EU":
+        input_file = PRIORITIZED_SCENARIO_GROUPS_EU_PATH
+    else:
+        raise ValueError(f"Dataset {dataset_name} not supported.")
 
-    # Define the scenario groups that are outside of the simulator's capabilities
+    # Load dataset
+    df = pd.read_excel(input_file)
+
+    # Scenarios to remove for all simulators
     scenarios_to_remove = ["Control Loss", "Human fault", "Animal Interaction", "Visibility"]
 
-    # Filter out the unwanted scenarios
+    # Filter
     df_filtered = df[~df['Scenario_Group'].isin(scenarios_to_remove)]
 
-    # Save the filtered DataFrame back to a new Excel file
-    output_file = f'filtered_scenarios_{simulator_name}.xlsx'
+    # Output path based on simulator selection
+    simulator_output_map = {
+        "Carla": FILTERED_SCENARIOS_CARLA_PATH,
+        "Gazebo": FILTERED_SCENARIOS_GAZEBO_PATH,
+        "Audacity": FILTERED_SCENARIOS_AUDACITY_PATH,
+        "LGSVL": FILTERED_SCENARIOS_LGSVL_PATH
+    }
+
+    if simulator_name not in simulator_output_map:
+        raise ValueError(f"Simulator {simulator_name} not supported.")
+
+    output_file = simulator_output_map[simulator_name]
+
+    # Save output
     df_filtered.to_excel(output_file, index=False)
 
-    print(f"scenarios are filter based on the selected simulator, see file {output_file}")
+    print(f"Scenarios filtered for {simulator_name}, saved to: {output_file}")
